requirements.txt
alembic==1.12.0
amqp==5.1.1
annotated-types==0.5.0
ansible==8.2.0
ansible-core==2.15.2
anyio==3.7.1
asgiref==3.5.2
async-timeout==4.0.3
asyncpg==0.28.0
autoflake==2.2.1
autopep8==1.7.0
bcrypt==4.0.1
billiard==4.1.0
black==23.9.1
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
Django==4.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==6.1.0
flower==2.0.1
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
humanize==4.8.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
pendulum==2.1.2
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.1
pyasn1==0.5.0
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.0.3
pydantic-settings==2.0.3
pydantic_core==2.3.0
pyflakes==3.1.0
PyMySQL==1.0.2
PyPDF2==3.0.1
pyright==1.1.330.post0
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
resolvelib==1.0.1
rsa==4.9
sentry-sdk==1.32.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.15.1
SQLAlchemy==2.0.21
sqlparse==0.4.2
starlette==0.27.0
toml==0.10.2
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.1
vine==5.0.0
wcwidth==0.2.8
WTForms==3.0.1
